requirements.txt
Django==5.1.5
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
psycopg2-binary==2.9.11
django-filter==24.3
dj-database-url==2.2.0
django-redis==5.4.0
boto3==1.35.0
django-storages==1.14.4
gunicorn==22.0.0
celery==5.4.0
django-celery-beat==2.7.0
django-celery-results==2.6.0
sentry-sdk==2.10.0
whitenoise==6.7.0
requests==2.32.3